+++ lib/bones_api_logging.dart
@@ -1,0 +1,6 @@
+/// Bones_API Logging Library.
+library bones_api_logging;
+
+export 'bones_api.dart';
+
+export 'src/bones_api_logging.dart';